Fix remaining instances of -Wshorten-64-to-32.
Updates a variety of types to reduce implicit casts; inserts explicit
casts where needed.

Bug: 879657
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3079134
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Kinuko Yasuda <kinuko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#911073}

+++ base/barrier_closure.h
@@ -20,7 +20,7 @@ namespace base {
 // the thread that calls the final Run() on the returned closures.
 //
 // |done_closure| is also cleared on the final calling thread.
-BASE_EXPORT RepeatingClosure BarrierClosure(int num_closures,
+BASE_EXPORT RepeatingClosure BarrierClosure(size_t num_closures,
                                             OnceClosure done_closure);
 
 }  // namespace base
